fix persist mock tests. update persist to require storage
